builderData input

+++ public/shaders/update.frag
@@ -1,7 +1,11 @@
 #version 300 es
-precision highp float;
-
-// out vec4 color;
+precision mediump float;
+// out vec3 newInfo;
+// out float aliveTime;
+// out vec4 Color;
 void main() 
 {
+    // Color = vec4(
+    // fract(newInfo.xy * 255.0),
+    // floor(newInfo.xy * 255.0) / 255.0);
 }
